+++ nodeconductor/cost_tracking/tests/factories.py
@@ -16,10 +16,12 @@ class PriceEstimateFactory(factory.DjangoModelFactory):
     total = factory.Iterator([10, 100, 1000, 10000, 980, 42])
     month = factory.Iterator(range(1, 13))
     year = factory.Iterator(range(2012, 2016))
+    limit = -1
 
     @classmethod
-    def get_list_url(self):
-        return 'http://testserver' + reverse('priceestimate-list')
+    def get_list_url(self, action=None):
+        url = 'http://testserver' + reverse('priceestimate-list')
+        return url if action is None else url + action + '/'
 
     @classmethod
     def get_url(self, price_estimate, action=None):