+++ nodeconductor/cost_tracking/views.py
@@ -1,5 +1,6 @@
 from __future__ import unicode_literals
 
+from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.models import Prefetch
@@ -7,6 +8,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from rest_framework import viewsets, exceptions, decorators, response, status
 
+from nodeconductor.core import views as core_views
 from nodeconductor.cost_tracking import models, serializers, filters
 from nodeconductor.structure import SupportedServices
 from nodeconductor.structure import models as structure_models
@@ -19,12 +21,13 @@ class PriceEditPermissionMixin(object):
         if self.request.user.is_staff:
             return True
         customer = reduce(getattr, scope.Permissions.customer_path.split('__'), scope)
-        if customer.has_user(self.request.user, structure_models.CustomerRole.OWNER):
+        is_owner = customer.has_user(self.request.user, structure_models.CustomerRole.OWNER)
+        if is_owner and settings.NODECONDUCTOR['OWNER_CAN_MODIFY_COST_LIMIT']:
             return True
         return False
 
 
-class PriceEstimateViewSet(PriceEditPermissionMixin, viewsets.ReadOnlyModelViewSet):
+class PriceEstimateViewSet(PriceEditPermissionMixin, core_views.ReadOnlyActionsViewSet):
     queryset = models.PriceEstimate.objects.all()
     serializer_class = serializers.PriceEstimateSerializer
     lookup_field = 'uuid'
@@ -34,13 +37,6 @@ class PriceEstimateViewSet(PriceEditPermissionMixin, viewsets.ReadOnlyModelViewS
         filters.PriceEstimateScopeFilterBackend,
         ScopeTypeFilterBackend,
     )
-
-    def get_serializer_class(self):
-        if self.action == 'threshold':
-            return serializers.PriceEstimateThresholdSerializer
-        elif self.action == 'limit':
-            return serializers.PriceEstimateLimitSerializer
-        return self.serializer_class
 
     def get_serializer_context(self):
         context = super(PriceEstimateViewSet, self).get_serializer_context()
@@ -109,6 +105,8 @@ class PriceEstimateViewSet(PriceEditPermissionMixin, viewsets.ReadOnlyModelViewS
         return response.Response({'detail': _('Threshold for price estimate is updated.')},
                                  status=status.HTTP_200_OK)
 
+    threshold_serializer_class = serializers.PriceEstimateThresholdSerializer
+
     @decorators.list_route(methods=['post'])
     def limit(self, request, **kwargs):
         """
@@ -146,6 +144,8 @@ class PriceEstimateViewSet(PriceEditPermissionMixin, viewsets.ReadOnlyModelViewS
         price_estimate.save(update_fields=['limit'])
         return response.Response({'detail': _('Limit for price estimate is updated.')},
                                  status=status.HTTP_200_OK)
+
+    limit_serializer_class = serializers.PriceEstimateLimitSerializer
 
 
 class PriceListItemViewSet(PriceEditPermissionMixin, viewsets.ModelViewSet):