+++ nodeconductor/server/base_settings.py
@@ -256,6 +256,7 @@ NODECONDUCTOR = {
     'CLOSED_ALERTS_LIFETIME': timedelta(weeks=1),
     'INVITATION_LIFETIME': timedelta(weeks=1),
     'OWNERS_CAN_MANAGE_OWNERS': True,
+    'OWNER_CAN_MODIFY_COST_LIMIT': True,
     'BACKEND_FIELDS_EDITABLE': True,
 }