+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Release NEXT
 - Allow to filter user by civil number.
 - Don't render superuser status. Drop unused viewsets.
 - Add LDAP scheme to service settings backend_url validator.
+- Add organization cost limit.
 
 Release 0.135.0
 ---------------